passing `word referring to itself` I've declared `currentlyParsedInputs` from the start and apparently never used it LOL. anyway, I discovered that when parsing, it's better to store the UDWs into its own "internal representation" first before evaluating, and this is what this commit does.
Previously, I evaluated the UDWs body at once and immediately put it
in the map (lookup). This causes a "loss of history" because at a
certain point, `foo` can't refer to itself because its value is
already overwritten in the lookup.

That's why it's better to store all of the UDWs body in another place
first, interpret the result, and store it in the lookup afterwards.